Rename cypher frame to be searchable
