+++ src/browser/modules/Stream/CypherFrame/CypherFrame.tsx
