+++ src/browser/modules/Stream/Stream.tsx
@@ -21,7 +21,7 @@
 import { connect } from 'react-redux'
 import React, { memo, useRef, useEffect } from 'react'
 import { StyledStream, Padding, AnimationContainer } from './styled'
-import CypherFrame from './CypherFrame/index'
+import CypherFrame from './CypherFrame/CypherFrame'
 import HistoryFrame from './HistoryFrame'
 import PlayFrame from './PlayFrame'
 import DefaultFrame from '../Frame/DefaultFrame'
